speechcoach-backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
whisper @ git+https://github.com/openai/whisper.git@main
google-generativeai==0.1.5
transformers==4.34.0
torch==2.0.1
corsheaders==3.15.0
